refactor: update auth file structure
